refactored scenes list to use proper react hooks
